refactor: 변경된 컴포넌트 경로 적용
