+++ pages/research/1.tsx
@@ -3,8 +3,8 @@ import { ChangeEvent, useState } from 'react';
 import { useRecoilState, useSetRecoilState } from 'recoil';
 import styled from 'styled-components';
 import Transitions from '~/layouts/Transitions';
-import { NextHead } from '~/components/common';
-import { Button, EstimateInput } from '~/components/domains';
+import { Button, NextHead } from '~/components/common';
+import { EstimateInput } from '~/components/research';
 import { initialResearch } from '~/utils/house';
 import { researchIndexState, researchState } from '~/atoms/research';
 import { PAGE_ROUTE } from '~/constants';
